Migrate install flow on webview api
