+++ src/html-utils.js
@@ -18,7 +18,6 @@ const logoLarge = fs.readFileSync(path.resolve(__dirname, '..', 'assets', 'image
 const proLogoSvg = fs.readFileSync(path.resolve(__dirname, '..', 'assets', 'images', 'kitepro.svg')).toString();
 const enterpriseLogoSvg = fs.readFileSync(path.resolve(__dirname, '..', 'assets', 'images', 'kiteenterprise.svg')).toString();
 const giftLogoPath = path.resolve(__dirname, '..', 'assets', 'images', 'icon-gift.png');
-const server = require('./server');
 
 const ASSETS_PATH = path.resolve(__dirname, '..', 'assets');
 
@@ -77,18 +76,6 @@ function wrapHTML (html) {
   html = handleInternalLinks(html);
   return `
   <style>
-    ${
-      process.env.NODE_ENV !== 'test'
-        ? `html {
-          font-size: ${vscode.workspace.getConfiguration('editor').get('fontSize')}px;
-        }
-
-        pre, code, .code {
-          font-family: ${vscode.workspace.getConfiguration('editor').get('fontFamily')};
-          font-size: ${vscode.workspace.getConfiguration('editor').get('fontSize')}px;
-        }`
-        : ''
-    }
     .icon-kite-gift::before {
       content: '';
       display: inline-block;
@@ -102,9 +89,6 @@ function wrapHTML (html) {
     }
   </style>
   ${STYLESHEETS}
-  <script>
-    window.PORT = ${server.PORT};
-  </script>
   ${SCRIPTS}
   <div class="kite platform-${os.platform()}">${handleInternalLinks(html)}</div>`
 }
